+++ src/main/java/kr/hhplus/be/server/common/config/SwaggerConfig.java
@@ -1,0 +1,18 @@
+package kr.hhplus.be.server.common.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("E-commerce API")
+                        .version("1.0.0")
+                        .description("이커머스 서비스 API 명세서"));
+    }
+}
